Wrapper component around horizontal visualization to prepare for future responsive vertical visualization
