+++ src/_scss/pages/account/overview/overview.scss
@@ -49,7 +49,7 @@
         color: $color-base;
         font-size: 14px;
         margin-top: 60px;
-        margin-bottom: 30px;
+        margin-bottom: 60px;
 
         @import "./sankey/sankey";
