+++ src/js/components/account/visualizations/sankey/components/SankeyMessage.jsx
@@ -1,0 +1,22 @@
+/**
+ * SankeyMessage.jsx
+ * Created by Kevin Li 4/18/17
+ */
+
+import React from 'react';
+
+const propTypes = {
+    message: React.PropTypes.string
+};
+
+export default class SankeyMessage extends React.Component {
+    render() {
+        return (
+            <div className="sankey-message">
+                {this.props.message}
+            </div>
+        );
+    }
+}
+
+SankeyMessage.propTypes = propTypes;
